Current

+++ system/scripts/system/sudo/_rotate_system_log.sh
@@ -2,7 +2,7 @@
 ContId=`docker inspect --format='{{.Id}}' $1`
 echo system container id $ContId >>/tmp/clean.log
 
-src=/var/lib/docker/containers/$system_cont_id/${ContId}-json.log 
+src=/var/lib/docker/containers/${ContId}/${ContId}-json.log 
 dest=/var/log/engines/raw/${ContId}-json.last
 
 if test -f $dest
@@ -10,6 +10,13 @@ if test -f $dest
  	rm $dest
  fi
 echo mv $src $dest  &>>/tmp/clean.log
-mv $src $dest  &>>/tmp/clean.log
-mv /var/log/engines/system_services/system/system.log /var/log/engines/system_services/system/system.log.last
+if test -f $src
+ then
+	cp $src $dest  &>>/tmp/clean.log
+	rm $src
+fi
+if test -f /var/log/engines/system_services/system/system.log
+ then
+	mv /var/log/engines/system_services/system/system.log /var/log/engines/system_services/system/system.log.last
+fi
 #exit
